requirements.txt
alabaster==0.7.10
autograd==1.1.11
babel==2.4.0
backports.weakref==1.0rc1
bleach==1.5.0
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
docutils==0.14
future==0.16.0
h5py==2.7.1
html5lib==0.9999999
idna==2.5
imagesize==0.7.1
imbalanced-learn==0.2.1
imblearn==0.0
jinja2==2.9.6
keras==2.0.6
markdown==2.6.8
markupsafe==1.0
matplotlib==2.0.0
munkres==1.0.12
networkx==1.11
nltk==3.2.2
pandas==0.19.2
pexpect==4.2.1
ptyprocess==0.5.2
pycorenlp==0.3.0
pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
pyyaml==3.12
regex==2016.12.27
retinasdk==1.0.0
scikit-learn==0.18.1
smart-open==1.5.3
snowballstemmer==1.2.1
sphinx==1.5.1
sphinxcontrib-websupport==1.0.1
tensorflow-gpu==1.2.1
tensorflow-tensorboard==0.1.2
theano==0.9.0
tqdm==4.11.1
typing==3.6.1
urllib3==1.22
werkzeug==0.12.2
